Create ACKNOWLEDGEMENT_template
We'll put this text in the documentation and at least one place in the code to be distributed.  Just including it here as a separate file so we have somewhere in the repo to copy it from.  We'll probably create a slightly edited copy, and maybe cut this down to just the relevant bits.