set up file types
Expanded class to include the notion of
file type.  A file class will be either a "real"
file or a "pad" file.  This has meaning when padding
out the parfu container.  This is done instead of
creating a separate class for pad files, since they
will basically have the same attributes, and this
way we can put pad files into vectors of regular files,
and some routines can use them or ignore them as
circumstances dictate.