part of spidering dir works
Pulled over legacy entry classification
function, but C++-ized it.  Had to mess around
with definition of functions and linking to make
it all work.

the spidering function is just a toy, but it works
in test mode and seems to traverse the target
directory correctly.  It doesn't DO anything, but
the machinery seems to run and it gets sane values
for file names and file sizes.