clean unkown type error message
When the spidering function encounters an entry
that it doesn't know what to do with (not a file,
not a symlink, not another directory) it sends
out a warning text message.

We may want to revise this after testing,
or perhaps make this a configuraable parameter.
We definitely want to warn about ignoring files,
if there are few of them, but if it's just a lot
of noise we'll want to do something else so
the output doesn't become a spew of spurious
error messages.