spider_dir possibly done
In spider_directory function, we're possibly now
correctly dealing with regularfiles, and symlinks,
and subdirectories.  Attempting to put together
traversing into subdirectories by making a recursive
call to have that subdirectory spider itself, which
I think is definitely the C++ way to do it.

We'll have to have some way to recognise loops in
directory trees, otherwise this will get tripped
up.  (Actually--the "is this spidered" boolean
may do that for us by mistake.  I don't know that
it will generate the correct behavior, but it should
prevent an infinite loop of recursion.