spider: readdir makes infinite loop
Due to whatever (which I can't right now be arsed
to sort out) readdir that this code gets at on Hydro
doesn't work right.  It doesn't return NULL at the
end of the directory, despite the man page saying
that's what will happen.

Preserving this version which just loops forever
on the same initial directory.  I'll go with
something else that will hopefully be sensible.