propogating sizes in target files
Getting target file and target file slice
classes filled out so that constructors are
available, to take the initialize sizes,
populating that in the directory spidering
function, and then starting (but not finishing)
the functions to calculate the positions of
files in the final container file.

This compiles, but haven't tested yet.

These functions will all be used to feed the
functions that create the catalog entries, since
they need to know the sizes and offsets and orders
of everything.