Parfu_storage_entry prelim implemented
Created class parfu_storage_entry.  File now a subclass.
hooks in to make files and symlinks as well.  Compiles
as is.

Had to declare the subclasses as "friend" classes to make
the initializtion work.  I guess the better way to do it
would be to have the subclass call the parent class's own
constructors, but may leave that as a future task.