created Parfu_storage_ref; updated target_collection
Created a Parfu_storage_ref which is a stuct that points to one of the
storage references, and carries with it the long in that is its size.
This is designed to be a constant-size (thus easily vectorizable)
reference to an entry.  The target collection is now two vectors
of this type, one for directories, and one for other entries.

Compiles, not tested.