debug cleanup and fix . and .. strip
Cleaned up a few of the debug printouts.

Added another one to debug why dot-files weren't being included.

I'd made a mistake in the comparison code.  I wasn't checking the
length, so anything that *began* with a "." was being excluded.  That
would have been a mildly annoying bug to find.  Seems to work now.
Found completely by accident by a typo in a test symlink.

Test code runs that includes the spidering directory class
function.  That seems to be working so far, including the recursive
descending-into-directory code.  It picks up subdirectories, symlinks,
and real files.