prelim root dir into target collection
I have a preliminary constructor to put a directory that's
the root of a spidered directory into a target_collection
class.  It builds, but only the first half of it
(the directories) but not the files are implemented.

Going to instantiate this with some debug code to make
sure that the subdirectory thing is working correctly.