Target constructor works, one layer
The target constructor class constructor works, sort of.  It
successfully creates the class, it populates it from the
root directory...but only the first layer.  I need to add
code so that it keeps going down through the directory
tree and adds all the directories, not just the ones on the
top layer.

However, within the functionality it has, this seems to
run correctly.