target_collection cstr works (just for dirs)
Lordy, this is a huge commit.  Sorry for anyone who
ever has to untangle this.

This started because the constructor function that I wrote
for the Parfu_target_collection class worked for the top
level of subdirectories worked but didn't work below that.
It consistently exited with a segfault.

So, turns out (pro tip!) that when you add to a vector, it
invalidates any iterator previously generated from that vector.
But (fun fact!) the iterator will continue to work within the
original confines of the vector (therwise I would have spotted
it much sooner).  (That last fact could well be implementation
dependent.)  So that's fun.  I spent a couple of hours chasing
pointer math and making sure all my references were valid.  Once
you run off the original vector the references the iterator
returns are invalid, but not nullptr.  So that was fun.

So the final fix was I modified the loop to iterate by just
comparing the iteration integer to size() rather than using
a formal iterator.

Along the way I discovered that I had mis-understood some of the
details about how class inheritance polymorphism worked.  So I had
to go back and fix a bunch of the function calls in the
storage_entry and directory classes, and change all of the function
calls to match.

Also also, I discoverd a mistake in how I was constructing filenames
for checking them for type within the directory
spidering function.  I have no idea how I missed that when
I was testing it.

Also also also, I fixed (I think) the return logic for the
legacy function that determines an entry type.  If the file
doesn't exist, it actually returns that as a return code, rather
than "invalid file type".  That's tough to test, because the
only time that would every be thrown would be if one of the
other functions had faulty logic, or if entries in the directory
that was being spidered are disappearing between being added to
the list and being checked for type.

This works JUST to bring in the subdirectories, so while this is
working as far as it goes, it is NOT complete and does not work
as it needs to.  I need to write the code that then adds the files
and symlinks, but those will be relatively easy.