fix to Parfu_target_collection (dironly)
The root directory that's been spidered alos
has to be in the directory list in case there are
top-level files.  The last edit set up the class
constructor to create that correctly, however, there
was a typo in the code.  This fixes that typo and
this is after the new version has been tested and
verified to work.