convert directory subfiles to ptr vec
Setting up for the rest of the storage target
collection.  Changing the internal representation
of files in the directory class to pointers.  That
makes them easier to port over from directories, since
they're that way.  That also potentially makes it easier
to have a collection of unrelated files in a big list so
you don't care how theyr'e arranged.

Builds, runs.