target_collection also has files
The Parfu_target_collection class is now also bringing
in files and symlinks as well as directories.  Builds,
runs, havne't tested yet if it does the right thing.

When implementing it, got a really weird
"is an inaccessible base of" error, which didn't make
any sense because I wasn't alling ANTHING from the
derived class it was mentioning.  But declaring that
class as a "public" subclass of its parent fixed it.

Again, compiles, runes, don't know if it's right.