target_collection has files and works
In the previous version, I left out a function
definition in the derived class for the number of
subfiles for directories it was always returning hard
zero.

So the Parfu_target_collection class constructor that takes
a root directory as its argument now compiles, runs, and as
far as small testing indicates, works correctly.