target_collection includes symlinks
The Parfu_target_collection class constructor now correctly
imports and includes symlinks.  The dump class functoin
correctly dumps them.  The "order size" for directories and
symlinks being set correctly.

And have now tested that non-zero-length files seem to be
correctly represented.  So the upshot of this is that we
can create a single, linear list of all the directories,
order them by size, and they will be in the correctly order
for splitting and for putting in order for archiving.