target_collection sets offsets
The Parfu_target_collection function set_offsets now works.
A bunch of changes and auxiliary function changes.  This
now lays out the locations of target files in the
archive file.

Dump functions also implemented and tested.  Tested
on hundreds of files; sorting and ordering and offsets
seem to be right.  Haven't specifically checked for
overlaps yet.