working on marching order function
I separated out the printing function from the
calculate and export marching order function.  I left
it in the storage class because that simplifies getting
at some of data from other classes even though it's not
strictly necessary to be htere and could be done by
other means with a better function interface.

Compiles, not complete, so not tested.