split print marching order function
Continuing to build up the main function in
target_collection that puts together the
buffers of rank marching orders.  I refactored
the print_marching_order function into two; one
that takes the internal offsets in the slices as
given, and the other "_raw" version that takes all
the sizes and offsets from call parameters, so that
those can be controlled from the logic that's
splitting files across rank buckets.