marching orders possibly compplete; cleanup
The function that prints out the buffers with the
per-rank marching orders may be complete.  The two
branches that 1) take care of files who can fit
within a bucket and 2) spill out over more than one
are both done.

Compile, not tested.

Also did a bit of cleanup, changing the type of the
slice class's "position in archive" variable to
unsigned long rather than signed.