marching orders: shifted push
In the function that creates the marching orders, I realized
that the push to create a new empty entry should be right before
appending to the last buffer rathr than after it, so that
no buffer would ever be left empty, in the middle but also
the last one.

Compiles, not tested.