several fixes in and tested
Did a test today and there were a few problems with
the marching orders coming out.  The type designations
were wrong.  The file position and archive positions were
messed up, and the header sizes were all -1 instead of
512 or something.

The position values were because I was calling a function
with the arguments in the wrong order.  Fixed.

There were several problems with the type designators.  I
went through and fixed them and put it all back together.
they're working correctly now.

The header size was a subtle one.  I had created copy
constructors for the Parfu_slice class.  But *later*,
I added the header size variable as a member, but didn't
update the assignment constructor.  So everything worked fine,
EXCEPT when you used a vector of Slices, and push() something
onto the back, it lost the header size value because of the
assignemtn operatoir.  ironically, the default one the compiler
built worked fine.  In other words, I commented out the
assignment operator I created and it suddenly worked right.

The code compiles and runs.  The marching order file generated
are all correct as far as I can see.  I neeed to make another
revision and get rid of all th edebugging output.