Added data move files
Added .cc and .hh files to hold the code to take
orders and move the data for archiving.

The .hh file contains a structure that will hold one single
move order (a line from an order file).  Next we'll make a
class that will manage a vector of these.  The idea is
that one order buffer will be rolled into one such class
instance.

This builds; there's nothing to test yet.

Makefile modified so that they are built.