file open correctly deadlocks
This is legit a good thing.

I'm building a prototype of the rank zero
parfu main.  It takes a directory, spiders it, and
then does all the stuff to prepare to take that list
as the target list.  It now also takes the name of
the output archive file.

It's doing the MPI_Init stuff, and only rank zero is
doing the file spidering stuff.  It sends the initial
order ("create") along with the name of the archive
file to all the ranks, then it needs to do a collective
file open of the archive file.

Since I haven't written the worker rank function yet,
rank zero sets up orders, broadcasts them, then tries
to run the collective open.  It correctly deadlocks
because I haven't written the bit where the worker
ranks also participate in the file open.