rank filename comm works
Basic bcast communication works from rank 0 to all other ranks.
Builds, has been tested.

Pack up an order which is a single letter and then the name of
the archive filename.  Send that broadcast from rank zero to other
ranks.  The other ranks get the order letter, get the filename
correctly.

Right now, rank zero's collective open is commented out, because
the worker ranks collective open isn't set up yet.  When
the collective is opened on 0 but not on the others, it
deadlocks.  Now that it's commented out on rank 0, everything
runs, prints out its debugging data, and then everyone exists.

So now to get the collective file open set up for all the
ranks.  Almost ready to move data.