prototype worker receive loop
Prelminary implementation of the state loop where the worker nodes
loop forever waiting for instructions.  It has two different
states, one for receiving broadcasts, and one for receiving
individual messages.  It starts in broadcast but can be flipped
by a message instruction.  Then an individual packet can flip
it back.

All the clauses are implemented except for the actual one that
will do the file movement.  That's next once the communication
is shown to be working, and rank 0 is shown to be able to shut
down the worker nodes by sending them the shutdown command.