iNdividual loop works, indv send too
Added some debug printouts to the worker thread functions for
debugging.  From test main, broadcast instruction message to
threads to switch to individual mode.  That worked.

Was able to them cycle through and send individual shutdown
message, and the result is that all threads exited cleanly.

Changed the send function to take a tag as a parameter.  Turns
out MPI_ANY_TAG is meant as a function for the receive call, but
does NOT work in the send call.  For now just ending all with
tag=0.

Works, tested, still has debugging output and lots of warnings
for unused variables as I roll things into utility functions.