collective open works!
Now that we've shuffled stuff off to auxiliary functions,
made a sweep to remove unused variables.  Copiles without
warnigns now.

Set up error checking for all the MPI calls.  The collective file
open was not working.  Turns out I'd put the wrong string in as
the file name.  Itw as trying to open the CWD as a file to write
to, which did not go well.

Fixed that now.  Runs with all MPI calls successfully completing.