excluseive file create; works
Cleaned up a few debug outputs.

Definitely verified that collective file open succeeds when the file
is valid, and when the file isn't.  In both cases output is informative.

Set the mode to exclusive, so that it will never overwrite an
archive file.  The user will need to get it out of the way first.