remove debug; working on boss funcs
Removed some debug cerr statements I used to debug one of the
problems recently.  (It was an off-by-one; I was writing to
rank 1 but forgot that I needed to transmit order set 0.)

Also getting the "transmit all th orders" function written.
It isn't being invoked yet, but everything compiles and runs
the way it used to.