fixed deadlock and msg error
Had an off-by-one deadlock error in the loop that waited for
all the ranks to report back.  Started on i=0 rather than
i=1, so waiting for one more rank than was ever going to
report back.  Fixed that.

Also had a problem with messaging the "I'me done" message.  I
mis-configured both the send and the receive in interesting ways.
Fixed.

Built, tested, runs as expected, with some extra debugging
output.