pad arc to blksz; now tar compat
Tar apparently is very persnickety about the size of the tarfile
being fully a multiple of the tar BLOCKSIZE.  It will complain if
it's not.

Now pad out every bucket with nulls to make them even BLOCKSIZE.

Now (so far) produces tar files that tar unpacks happily and doesn't
complain about.