bucket size mult of 512
The bucket size must (apparentlY) be a multiple of the
tar blocksize (512 bytes).  This doesn't matter when the
files are below bucket size, but does matter when they're
bigger.  If it's not, the code that pads out the last bucket
to the nearest bucket size wanders off the end, or something.

Compiles, works, is now producing archive files of correct
size again.