check for block size
the _Create function now has a last-minute check to make sure the
bucket size is a multiple of the tar block size.  It won't save
the code but it will at least warn the user.

Future user-supplied block sizes will have to be adjusted to the
nearest multiple of that block size.