clean debug + added max_orders parameter
Cleaned up some debugging that I forgot to check in.  Debug
prints now down to roughly what I want.  Also, now have printouts
of the counts of the per-bucket order bundles that are going out.
In that process I discovered I was sending double of the last
order, I think, so that's now fixed.

I also changed the function call for the function that splits
up the whole order list into bucket bundles.  There's now a
parameter that (when implemented) will set a maximum number
of file entries per bundle, which when tied to a command-line
parameter, will help the user tune the efficiency of small
files vs. larger files.

Builds, runs, output as expected.