has limit of # orders per bucket
I've implemented a parameter to the "split orders" function that
allows input of a max number of orders per bucket.  That allows
parfu to spread the pain of metata load across more ranks, rather
than having it concenetrated on one or a couple of ranks.

The debugging output that lists number of orders per bucket bundle
and total size of the bucket orders will be a very useful
output to tune against.