argument parser almost complete
Composed an argument parser function.  That's working;
it's correctly pulling out the argument that sets the
max number of orders per bundle, and that is now used
as the actual maximum.

It's also processing the non-flagged arguments as targets

That all builds and runs correctly.  It's not hooked into
the main code yet.  I still need to add the flag for
"target file", and then add flags for bucket size and
archive file multiplier.