usage(), CL params work, tested
Command-line parsing seems to work in all cases.
Max-orders-per-bucket and bucket-size tested to go into
the code and work as expected.  Possibly still needs a bit
of error reporting turned down, but it's now in the realm of
reasonable for a not-1.0 code.

I will probably soon declare this version 0.6, and use that for
all the performance and verification testing.  Its command
structure is enough for that.